llvm: make clear distinstion between points-to analysis and PSS builder
when we wanted to get the llvm->PSSNode mapping, we needed to
use the builder. This is not very nice (I need points-to information,
why I'm calling methods on builder instead of some PSS or PTSet?).
So proxy the builder via LLVMPointsToAnalysis class.
LLVMPointsToAnalysis is now the interface that creates the
pointer state subgraph and then, using real implementation of the PSS
analysis, computes the points-to information

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>