add recursive2 test
the test calculates factorial using simple recursive functions.
current block mapping has a problem with recursive functions:
recursive function's artificial return node (NOOP)
does not have a block assigned during function construction.
so if there is a return to some block B, this block's predecessor needs
to be block of the NOOP node, which is not known at that time.

from now on, block addition is attempted with every new return, because
the block is already known after the whole function is constructed.

Signed-off-by: Tomas Jasek <tomsik68@gmail.com>